requirements.txt
fastapi
uvicorn
openai
python-dotenv
pinecone-client
chromadb
nltk
cohere
rapidfuzz
numpy
